# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: Payment Gateway API - Enterprise Payment Processing
  description: |
    # 🚀 Payment Gateway API v1.0
    
    **Enterprise-grade payment processing API** with comprehensive Authorize.Net integration, designed for secure, scalable, and compliant financial transactions.
    
    ## 🎯 Core Capabilities
    
    ### 💳 Payment Processing
    - **Direct Purchase**: One-step authorization and capture
    - **Two-Step Processing**: Separate authorization and capture
    - **Transaction Management**: Void and refund operations
    - **Payment Method Validation**: Pre-transaction verification
    
    ### 📄 Subscription Management
    - **Recurring Billing**: Automated subscription processing
    - **Plan Management**: Flexible subscription plans with trials
    - **Lifecycle Control**: Pause, resume, and cancel subscriptions
    - **Dunning Management**: Intelligent failed payment handling
    
    ### 📊 Analytics & Reporting
    - **Real-time Dashboards**: Live transaction monitoring
    - **Business Intelligence**: Revenue and performance analytics
    - **Compliance Reports**: PCI DSS, GDPR, and SOX reporting
    - **Data Export**: Comprehensive data export capabilities
    
    ### 🔔 Webhook Integration
    - **Real-time Notifications**: Instant transaction updates
    - **Event Processing**: Comprehensive event handling
    - **Retry Logic**: Guaranteed delivery with exponential backoff
    
    ## 🔐 Security & Authentication
    
...
    
...
    
...
    
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /subscriptions:
    post:
      tags:
        - Subscription Management
      summary: Create a new subscription
      description: |
        Create a new subscription for recurring billing. Supports trial periods,
        custom billing cycles, and various subscription plans.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '201':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags:
        - Subscription Management
      summary: List subscriptions
      description: Retrieve a list of subscriptions with filtering and pagination.
      parameters:
        - name: customerId
          in: query
          description: Filter by customer ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter by subscription status
          schema:
            $ref: '#/components/schemas/SubscriptionStatus'
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Subscriptions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionListResponse'
...
  /analytics/dashboard:
    post:
      tags:
        - Analytics
      summary: Get dashboard metrics
      description: |
        Retrieve comprehensive dashboard metrics including transaction volumes,
        revenue data, and performance indicators for a specified time period.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardRequest'
      responses:
        '200':
          description: Dashboard metrics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardMetrics'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /analytics/revenue:
    get:
      tags:
        - Analytics
      summary: Get revenue analytics
      description: Retrieve revenue analytics including MRR, churn rates, and growth metrics.
      security:
        - BearerAuth: []
      parameters:
        - name: startDate
          in: query
          required: true
          description: Start date for analysis (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          required: true
          description: End date for analysis (ISO 8601)
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Revenue analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevenueMetrics'

  # Webhook Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
...
  schemas:
...
    CreateSubscriptionRequest:
      type: object
      required:
        - customerId
        - planId
        - paymentMethodId
      properties:
        customerId:
          type: string
          format: uuid
        planId:
          type: string
          format: uuid
        paymentMethodId:
          type: string
          format: uuid
        trialPeriodDays:
          type: integer
          minimum: 0
          example: 14
        metadata:
          type: object
          additionalProperties: true
...
    SubscriptionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        subscriptionId:
          type: string
          example: sub_1234567890
        customerId:
          type: string
          format: uuid
        planId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/SubscriptionStatus'
        currentPeriodStart:
          type: string
          format: date-time
        currentPeriodEnd:
          type: string
          format: date-time
        nextBillingDate:
          type: string
          format: date-time
        trialStart:
          type: string
          format: date-time
        trialEnd:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SubscriptionStatus:
      type: string
      enum:
        - ACTIVE
        - PAUSED
        - CANCELLED
        - EXPIRED
        - PAST_DUE
        - PENDING
    SubscriptionListResponse:
      type: object
      properties:
        subscriptions:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8

    # Analytics Schemas
    DashboardRequest:
      type: object
      required:
        - startDate
        - endDate
      properties:
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        includeComparison:
          type: boolean
          default: false
    DashboardMetrics:
      type: object
      properties:
        transactionMetrics:
          $ref: '#/components/schemas/TransactionMetrics'
        revenueMetrics:
          $ref: '#/components/schemas/RevenueMetrics'
        subscriptionMetrics:
          $ref: '#/components/schemas/SubscriptionMetrics'
        performanceMetrics:
          $ref: '#/components/schemas/PerformanceMetrics'
    TransactionMetrics:
      type: object
      properties:
        totalTransactions:
          type: integer
          example: 1250
        successfulTransactions:
          type: integer
          example: 1200
        failedTransactions:
          type: integer
          example: 50
        successRate:
          type: number
          format: decimal
          example: 96.0
        averageTransactionValue:
          type: number
          format: decimal
          example: 89.50
    RevenueMetrics:
      type: object
      properties:
        totalRevenue:
          type: number
          format: decimal
          example: 125000.00
        recurringRevenue:
          type: number
          format: decimal
          example: 45000.00
        refundedAmount:
          type: number
          format: decimal
          example: 2500.00
        netRevenue:
          type: number
          format: decimal
          example: 122500.00

    # Webhook Schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error category
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
          example: Invalid payment method
        details:
          type: string
          description: Detailed error information
        correlationId:
          type: string
          description: Request correlation ID for tracing
          example: corr-123e4567-e89b-12d3-a456-426614174000
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: API endpoint path
          example: /api/v1/payments/purchase
...
  responses:
    BadRequest:
      description: Bad request - invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: VALIDATION_ERROR
            message: Invalid request parameters
            details: Amount must be greater than 0
            correlationId: corr-123e4567-e89b-12d3-a456-426614174000
            timestamp: "2025-09-10T14:30:00Z"
            path: /api/v1/payments/purchase
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: AUTHENTICATION_ERROR
            message: Invalid or expired token
            correlationId: corr-123e4567-e89b-12d3-a456-426614174000
            timestamp: "2025-09-10T14:30:00Z"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: AUTHORIZATION_ERROR
            message: Insufficient permissions to access this resource
            correlationId: corr-123e4567-e89b-12d3-a456-426614174000
            timestamp: "2025-09-10T14:30:00Z"
...